seo-blog-gen-app/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
transformers==4.30.2
torch==2.0.1
requests==2.31.0
